+++ src/Controller/PersonneController.php
@@ -3,7 +3,8 @@
 namespace App\Controller;
 
 use App\Entity\Personne;
-use App\Form\PersonneType;
+use App\Form\PersonneNewType;
+use App\Form\PersonneEditType;
 use App\Repository\PersonneRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -11,8 +12,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 class PersonneController extends AbstractController
-{
-    
+{   
     public function index(PersonneRepository $personneRepository): Response
     {
         return $this->render('personne/index.html.twig', [
@@ -23,15 +23,16 @@ class PersonneController extends AbstractController
 
     public function new(Request $request): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_SECRETAIRE');
         $personne = new Personne();
-        $form = $this->createForm(PersonneType::class, $personne);
+        $form = $this->createForm(PersonneNewType::class, $personne);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($personne);
             $entityManager->flush();
-
+            $this->addFlash('success', 'Personne créée !');
             return $this->redirectToRoute('personne_index');
         }
 
@@ -50,12 +51,13 @@ class PersonneController extends AbstractController
 
     public function edit(Request $request, Personne $personne): Response
     {
-        $form = $this->createForm(PersonneType::class, $personne);
+        $this->denyAccessUnlessGranted('ROLE_SECRETAIRE');
+        $form = $this->createForm(PersonneEditType::class, $personne);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $this->getDoctrine()->getManager()->flush();
-
+            $this->addFlash('success', 'Personne modifiée !');
             return $this->redirectToRoute('personne_index');
         }
 
@@ -67,10 +69,12 @@ class PersonneController extends AbstractController
 
     public function delete(Request $request, Personne $personne): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
         if ($this->isCsrfTokenValid('delete'.$personne->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($personne);
             $entityManager->flush();
+            $this->addFlash('success', 'Personne supprimée !');
         }
 
         return $this->redirectToRoute('personne_index');
